Header chain proofs in batches (#659)
* feat: Add header_chain_proof_batch_size to protocol params.

* chore: Apply new clippy suggestions.

* feat(database/hcp): Add get_latest_block_height.

* feat: Add is_batch_ready and prove_blocks.

* fix: Return option in get_next_non_proven_block.

* feat(database): Add get_latest_proven_block_info.

* feat: Add get_next_n_non_proven_block db func.

+++ scripts/schema.sql
@@ -28,7 +28,7 @@ create table if not exists header_chain_proofs (
     block_hash text primary key not null,
     block_header text,
     prev_block_hash text,
-    height int not null,
+    height bigint not null,
     proof bytea
 );
 create table if not exists watchtower_xonly_public_keys (
